pull

+++ core/mybatis-generator-systests-ibatis2-java2/src/test/java/mbg/test/ib2j2/conditional/CountByExampleTest.java
@@ -1,3 +1,18 @@
+/**
+ *    Copyright 2006-2016 the original author or authors.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
 package mbg.test.ib2j2.conditional;
 
 import mbg.test.common.util.TestUtilities;
@@ -50,7 +65,7 @@ public class CountByExampleTest extends AbstractConditionalJava2Test {
             FieldsonlyExample example = new FieldsonlyExample();
             example.createCriteria().andIntegerfieldGreaterThan(new Integer(5));
     
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
     
             example.clear();
@@ -82,7 +97,7 @@ public class CountByExampleTest extends AbstractConditionalJava2Test {
     
             PkonlyExample example = new PkonlyExample();
             example.createCriteria().andIdGreaterThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
     
             example.clear();
@@ -115,7 +130,7 @@ public class CountByExampleTest extends AbstractConditionalJava2Test {
     
             PkfieldsblobsExample example = new PkfieldsblobsExample();
             example.createCriteria().andId1NotEqualTo(new Integer(3));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
     
             example.clear();
@@ -147,7 +162,7 @@ public class CountByExampleTest extends AbstractConditionalJava2Test {
     
             PkfieldsExample example = new PkfieldsExample();
             example.createCriteria().andLastnameLike("J%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
     
             example.clear();
@@ -176,7 +191,7 @@ public class CountByExampleTest extends AbstractConditionalJava2Test {
     
             PkblobsExample example = new PkblobsExample();
             example.createCriteria().andIdLessThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
     
             example.clear();
@@ -207,7 +222,7 @@ public class CountByExampleTest extends AbstractConditionalJava2Test {
     
             FieldsblobsExample example = new FieldsblobsExample();
             example.createCriteria().andFirstnameLike("S%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
     
             example.clear();
@@ -254,7 +269,7 @@ public class CountByExampleTest extends AbstractConditionalJava2Test {
     
             AwfulTableExample example = new AwfulTableExample();
             example.createCriteria().andEMailLike("fred@%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
     
             example.clear();
