+++ core/mybatis-generator-systests-ibatis2-java2/src/test/java/mbg/test/ib2j2/hierarchical/HierarchicalJava2Test.java
@@ -1,19 +1,18 @@
-/*
- *  Copyright 2006 The Apache Software Foundation
+/**
+ *    Copyright 2006-2016 the original author or authors.
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
  */
-
 package mbg.test.ib2j2.hierarchical;
 
 import java.math.BigDecimal;
@@ -25,11 +24,11 @@ import java.util.List;
 import mbg.test.common.util.TestUtilities;
 import mbg.test.ib2j2.generated.hierarchical.dao.AwfulTableDAO;
 import mbg.test.ib2j2.generated.hierarchical.dao.FieldsblobsDAO;
+import mbg.test.ib2j2.generated.hierarchical.dao.FieldsonlyDAO;
 import mbg.test.ib2j2.generated.hierarchical.dao.PkblobsDAO;
 import mbg.test.ib2j2.generated.hierarchical.dao.PkfieldsDAO;
 import mbg.test.ib2j2.generated.hierarchical.dao.PkfieldsblobsDAO;
 import mbg.test.ib2j2.generated.hierarchical.dao.PkonlyDAO;
-import mbg.test.ib2j2.generated.hierarchical.dao.subpackage.FieldsonlyDAO;
 import mbg.test.ib2j2.generated.hierarchical.model.AwfulTable;
 import mbg.test.ib2j2.generated.hierarchical.model.AwfulTableExample;
 import mbg.test.ib2j2.generated.hierarchical.model.AwfulTableKey;
@@ -212,7 +211,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             FieldsonlyExample example = new FieldsonlyExample();
             example.createCriteria().andIntegerfieldGreaterThan(new Integer(5));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -384,7 +383,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkonlyExample example = new PkonlyExample();
             example.createCriteria().andIdGreaterThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example = new PkonlyExample();
@@ -1097,7 +1096,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkfieldsExample example = new PkfieldsExample();
             example.createCriteria().andLastnameLike("J%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1387,7 +1386,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkblobsExample example = new PkblobsExample();
             example.createCriteria().andIdLessThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1772,7 +1771,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             PkfieldsblobsExample example = new PkfieldsblobsExample();
             example.createCriteria().andId1NotEqualTo(new Integer(3));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1959,7 +1958,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             FieldsblobsExample example = new FieldsblobsExample();
             example.createCriteria().andFirstnameLike("S%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -2977,7 +2976,7 @@ public class HierarchicalJava2Test extends AbstractHierarchicalJava2Test {
 
             AwfulTableExample example = new AwfulTableExample();
             example.createCriteria().andEMailLike("fred@%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
